Game theoretically done

+++ Connect 4/timer.h
@@ -32,7 +32,7 @@ public:
             endTime = m_EndTime;
         }
 
-        return std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count();
+        return double(std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count());
     }
 
     double elapsed_seconds()
